+++ react/components/EXPERIMENTAL_Select/colors.js
@@ -3,6 +3,10 @@ export default {
   blue: '#134cd8',
   gray: '#979899',
   lightGray: '#f2f4f5',
-  'muted-4': '#E3E4E6',
+  'c-on-base': '#3f3f40',
+  'muted-1': '#727273',
+  'muted-2': '#979899',
+  'muted-3': '#cacbcc',
+  'muted-4': '#e3e4e6',
   red: '#ff4c4c',
 }