+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `FilterBar` clear icon disabled style.
 - `AutocompleteInput` disabled style.
 - Filters hidden for existing statements.
+- `EXPERIMENTAL_Select` style.
 
 ### Added